+++ nodes/homee-device.js
@@ -10,16 +10,22 @@ module.exports = function (RED) {
     this.name = config.name;
     this.nodeId = parseInt(config.nodeId, 10);
     this.profile = parseInt(config.profile, 10);
+    this.statusTemplate = config.statusTemplate;
     this.storageConfigured = RED.settings.contextStorage && 'homeeStore' in RED.settings.contextStorage;
 
-    if (this.nodeId === -1) throw new Error('The node id must not be -1');
+    if (this.nodeId === -1) throw new Error(RED._('homeeDevice.error.homee-node-id'));
 
     try {
-      this.attributes = JSON.parse(config.attributes);
-      if (!Array.isArray(this.attributes)) throw new Error('Attributes must be an array');
+      if (typeof config.attributes === 'string') {
+        this.attributes = JSON.parse(config.attributes);
+      } else {
+        this.attributes = config.attributes;
+      }
+
+      if (!Array.isArray(this.attributes)) throw new Error(RED._('homeeDevice.error.attributes-must-be-array'));
 
       if (this.attributes.filter((a) => a.node_id !== this.nodeId).length) {
-        throw new Error('The node id of at least one attribute does not match the device node id');
+        throw new Error(RED._('homeeDevice.error.node-ids-dont-match'));
       }
 
       if (this.storageConfigured) {
@@ -31,10 +37,11 @@ module.exports = function (RED) {
 
           attributes.forEach((storedAttribute) => {
             const attribute = this.attributes.find((a) => a.id === storedAttribute.id);
+            if (!attribute) return;
             attribute.current_value = storedAttribute.current_value;
             attribute.target_value = storedAttribute.target_value;
             attribute.data = storedAttribute.data;
-            
+
             this.virtualHomeeNode.api.send(JSON.stringify({ attribute }));
           });
 
@@ -43,7 +50,7 @@ module.exports = function (RED) {
       }
 
       this.device = new Device(this.name, this.nodeId, this.profile, this.attributes, this.icon);
-      this.status({ fill: 'green', shape: 'dot', text: this.device.statusString() });
+      this.status({ fill: 'green', shape: 'dot', text: this.device.statusString(this.statusTemplate) });
 
       this.virtualHomeeNode.registerDevice(this.id, this.device, (err) => {
         if (err) throw Error(err);
@@ -56,7 +63,7 @@ module.exports = function (RED) {
     // new value from flow
     this.on('input', (msg) => {
       if (typeof msg.payload !== 'object') {
-        node.warn('Only JSON-Objects are valid payloads. Ignoring message.');
+        node.warn(RED._('homeeDevice.warning.invalid-payload'));
         return;
       }
 
@@ -70,7 +77,11 @@ module.exports = function (RED) {
       Object.keys(msg.payload).forEach((key) => {
         switch (key) {
           case 'attribute':
-            this.updateAttribute(msg.payload.attribute.id, msg.payload.attribute.value, msg.payload.attribute.data);
+            this.updateAttribute(
+              msg.payload.attribute.id,
+              msg.payload.attribute.value,
+              msg.payload.attribute.data,
+            );
             break;
           case 'attributes':
             msg.payload.attributes.forEach((a) => this.updateAttribute(a.id, a.value, a.data));
@@ -94,7 +105,7 @@ module.exports = function (RED) {
         if (err) node.debug(`Can't store data for device #${this.nodeId}, ${err}`);
 
         node.debug(`stored data for device #${this.nodeId}`);
-        node.status({ fill: 'red', shape: 'dot', text: this.device.statusString() });
+        node.status({ fill: 'red', shape: 'dot', text: this.device.statusString(this.statusTemplate) });
         done();
       });
     });
@@ -114,14 +125,14 @@ module.exports = function (RED) {
 
     /**
      * update attribute
-     * @param  {int} id        the attribute id
+     * @param  {int} id  the attribute id
      * @param  {int|float} value  new value
      * @param  {string} data    new data
      * @return {void}
      */
     this.updateAttribute = (id, value, data) => {
       if (typeof id !== 'number' || typeof value !== 'number') {
-        node.warn('id and value must be numeric. ignoring message.');
+        node.warn(RED._('homeeDevice.warning.numeric-id-value'));
         return;
       }
 
@@ -129,15 +140,14 @@ module.exports = function (RED) {
       const unixTimestamp = Math.round(Date.now() / 1000);
 
       if (!attribute) {
-        node.warn(`Can't find attribute with id ${id}`);
+        node.warn(`${RED._('homeeDevice.warning.attribute-not-found')} ${id}`);
         return;
       }
 
       node.debug(`updating attribute #${id} to value: ${value}`);
 
       if (value < attribute.minimum || value > attribute.maximum) {
-        node.warn(`can't update attribute. The provided value must be
-            between ${attribute.minimum} and ${attribute.maximum}`);
+        node.warn(`${RED._('homeeDevice.warning.value-exceeds-min-max')} ${attribute.minimum} ${RED._('homeeDevice.warning.and')} ${attribute.maximum}`);
         return;
       }
 
@@ -148,18 +158,18 @@ module.exports = function (RED) {
       // first update target value only
       attribute.target_value = value;
       this.virtualHomeeNode.api.send(JSON.stringify({ attribute }));
-      
-      //save data if it is set
-      if (typeof data !== "undefined") {
+
+      // save data if it is set
+      if (typeof data !== 'undefined') {
         attribute.data = data;
-        }
+      }
 
       // next update current_value
       attribute.last_value = attribute.current_value;
       attribute.current_value = value;
       attribute.last_changed = unixTimestamp;
       this.virtualHomeeNode.api.send(JSON.stringify({ attribute }));
-      this.status({ fill: 'green', shape: 'dot', text: this.device.statusString() });
+      this.status({ fill: 'green', shape: 'dot', text: this.device.statusString(this.statusTemplate) });
     };
   }