+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.7.0] -
+### Added
+- Added german translation
+- homeeApi node: added filter options for outgoing messages
+- homeeDevice node: allow a custom template for the nodes status text
+- homeeDevice node: New attribute list for easier attribute definition
+- homeeDevice node: Attribute templates
+### Changed
+- Fixed: For attribute types FirmwareRevision, SoftwareRevision and DeviceProgramm the data attribute is used instead of the current value.
+- fixed icon size and switched from png to svg
+- homeeApi node: Changed icon for configuration node
+
 ## [0.6.0] - 2020-06-04
 ### Added
 - Added relationships (node <--> group) to global context